src/test/java/com/blogspot/ostas/leetcode/all/easy/third_maximum_number/SolutionTest.java
package com.blogspot.ostas.leetcode.all.easy.third_maximum_number;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class SolutionTest {

    /*
      Example 1:
    Input: nums = [3,2,1]
    Output: 1
    Explanation:
    The first distinct maximum is 3.
    The second distinct maximum is 2.
    The third distinct maximum is 1.
    Example 2:
    Input: nums = [1,2]
    Output: 2
    Explanation:
    The first distinct maximum is 2.
    The second distinct maximum is 1.
    The third distinct maximum does not exist, so the maximum (2) is returned instead.
    Example 3:
    Input: nums = [2,2,3,1]
    Output: 1
    Explanation:
    The first distinct maximum is 3.
    The second distinct maximum is 2 (both 2's are counted together since they have the same value).
    The third distinct maximum is 1.
      Constraints:
    1 <= nums.length <= 104
    -231 <= nums[i] <= 231 - 1
      Follow up: Can you find an O(n) solution?
    */
    @Test
    void example_0() {
        var solution = new Solution();
        int[] nums = new int[]{3, 2, 1};
        int expected = 1;
        var result = solution.thirdMax(nums);
        assertThat(result).isEqualTo(expected);
    }

    @Test
    void example_1() {
        var solution = new Solution();
        int[] nums = new int[]{1, 2};
        int expected = 2;
        var result = solution.thirdMax(nums);
        assertThat(result).isEqualTo(expected);
    }

    @Test
    void example_2() {
        var solution = new Solution();
        int[] nums = new int[]{2, 2, 3, 1};
        int expected = 1;
        var result = solution.thirdMax(nums);
        assertThat(result).isEqualTo(expected);
    }

}
